# Align SYCL `batched_gemv` to CUDA's