## Proposed changes

Align SYCL `batched_gemv[NT]` kernels to the CUDA variants. Resulted in ~3x faster performance overall in benchmark testing: `src/Platforms/tests/benchmark_accelBLAS "AccelBLAS gemv_batched benchmark"` on Aurora.

- Before:
    ```
    -------------------------------------------------------------------------------
    AccelBLAS gemv_batched benchmark
      gemv_batched<PlatformKind::SYCL>
    -------------------------------------------------------------------------------
    /home/leey/project/qmcpack/src/Platforms/tests/benchmark_AccelBLAS.cpp:407
    ...............................................................................

    benchmark name                       samples       iterations    estimated
                                         mean          low mean      high mean
                                         std dev       low std dev   high std dev
    -------------------------------------------------------------------------------
    [SYCL/f32]
    gemv_batched_137x79_N_bs64                     100             1    983.312 ms
                                            9.83662 ms    9.83348 ms    9.84051 ms
                                            17.7988 us    14.3506 us    24.5974 us

    [SYCL/f64]
    gemv_batched_137x79_N_bs64                     100             1     11.0358 s
                                            9.90384 ms    9.89675 ms    9.91111 ms
                                            36.7607 us    32.4087 us    44.3481 us

    [SYCL/f32]
    gemv_batched_137x79_T_bs64                     100             1    561.906 ms
                                            5.62591 ms    5.62261 ms    5.62947 ms
                                            17.4847 us    15.4053 us    22.1497 us

    [SYCL/f64]
    gemv_batched_137x79_T_bs64                     100             1    630.543 ms
                                            6.32057 ms    6.31739 ms    6.32379 ms
                                            16.3545 us    14.3819 us    20.5412 us
    ```
- After:
    ```
    -------------------------------------------------------------------------------
    AccelBLAS gemv_batched benchmark
      gemv_batched<PlatformKind::SYCL>
    -------------------------------------------------------------------------------
    /home/leey/project/qmcpack/src/Platforms/tests/benchmark_AccelBLAS.cpp:407
    ...............................................................................

    benchmark name                       samples       iterations    estimated
                                         mean          low mean      high mean
                                         std dev       low std dev   high std dev
    -------------------------------------------------------------------------------
    [SYCL/f32]
    gemv_batched_137x79_N_bs64                     100             1     12.1795 s
                                            2.65468 ms    2.65155 ms    2.65817 ms
                                            16.8654 us    14.4431 us    20.3502 us

    [SYCL/f64]
    gemv_batched_137x79_N_bs64                     100             1     14.8434 s
                                            3.16175 ms    3.15747 ms    3.16659 ms
                                              23.09 us    19.1353 us    32.2427 us

    [SYCL/f32]
    gemv_batched_137x79_T_bs64                     100             1    214.568 ms
                                            2.13945 ms    2.13535 ms    2.14369 ms
                                            21.1288 us    19.0325 us    23.9365 us

    [SYCL/f64]
    gemv_batched_137x79_T_bs64                     100             1    293.858 ms
                                             2.9381 ms    2.93492 ms    2.94142 ms
                                            16.6514 us      14.71 us    19.2061 us
    ```

## What type(s) of changes does this code introduce?

- Refactoring (no functional changes, no api changes)

### Does this introduce a breaking change?

- No

## What systems has this change been tested on?

Aurora

## Checklist

* * [x] I have read the pull request guidance and develop docs
* * [x] This PR is up to date with the current state of 'develop'
* * [x] Code added or changed in the PR has been clang-formatted
* * [ ] This PR adds tests to cover any new code, or to catch a bug that is being fixed
* * [ ] Documentation has been added (if appropriate)