@prckent 2025-08-27
Unfortunately we don't have any SYCL CI.

Before merging

- @ylee88 can you confirm that all the tests are OK?
- @ye-luo Any concerns?