@prckent 2025-09-05
test_one_gemv gets it with a batch size of 2 via test_AccelBlas.